requirements.txt
flask
flask-cors
pandas
scikit-learn
spacy
pdfplumber
python-docx
google-generativeai
python-dotenv